Add files via upload
You must have used the integer data type (int) in C++ by now, but do you know there is a
limit to how large or small a number can be stored in 'int'? The size of int is 4 bytes by
default (or 32 bits), which means the maximum number that can be stored in an integer is
2147483647 (231

-1), and the minimum number is -2147483648 (-2
31
-1). If you want to
store a larger or smaller integer, you can use 'long int' with a size of 8 bytes, but it will also
have a limited range. In this assignment, you will implement a class 'BigInteger' that will
allow you to store integers without any range limitations.